Refactor to use generalised controllers
