+++ jobs-api-server/src/main/java/org/zowe/jobs/services/JobsServiceV1.java
@@ -7,7 +7,7 @@
  *
  * Copyright IBM Corporation 2018, 2019
  */
-package org.zowe.jobs.v1.services;
+package org.zowe.jobs.services;
 
 import org.zowe.api.common.exceptions.ZoweApiException;
 import org.zowe.api.common.model.ItemsWrapper;
