+++ jobs-api-server/src/main/java/org/zowe/jobs/services/zosmf/GetJobsZosmfRequestRunner.java
@@ -7,7 +7,7 @@
  *
  * Copyright IBM Corporation 2019
  */
-package org.zowe.jobs.v1.services.zosmf;
+package org.zowe.jobs.services.zosmf;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
@@ -16,7 +16,7 @@ import lombok.extern.slf4j.Slf4j;
 
 import org.apache.http.HttpStatus;
 import org.apache.http.client.methods.RequestBuilder;
-import org.zowe.api.common.connectors.zosmf.ZosmfConnectorV1;
+import org.zowe.api.common.connectors.zosmf.ZosmfConnector;
 import org.zowe.api.common.exceptions.ZoweApiRestException;
 import org.zowe.api.common.model.ItemsWrapper;
 import org.zowe.api.common.utils.ResponseCache;
@@ -50,7 +50,7 @@ public class GetJobsZosmfRequestRunner extends AbstractZosmfJobsRequestRunner<It
     }
 
     @Override
-    protected RequestBuilder prepareQuery(ZosmfConnectorV1 zosmfConnector) throws URISyntaxException {
+    protected RequestBuilder prepareQuery(ZosmfConnector zosmfConnector) throws URISyntaxException {
         if (prefix == null) {
             prefix = "*";
         }
