+++ jobs-api-server/src/main/java/org/zowe/jobs/services/zosmf/ZosmfJobsServiceV1.java
@@ -7,13 +7,13 @@
  *
  * Copyright IBM Corporation 2018, 2020
  */
-package org.zowe.jobs.v2.services.zosmf;
+package org.zowe.jobs.services.zosmf;
 
 import lombok.extern.slf4j.Slf4j;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.zowe.api.common.connectors.zosmf.ZosmfConnectorV2;
+import org.zowe.api.common.connectors.zosmf.ZosmfConnector;
 import org.zowe.api.common.exceptions.ZoweApiException;
 import org.zowe.api.common.model.ItemsWrapper;
 import org.zowe.jobs.exceptions.JobFileIdNotFoundException;
@@ -22,14 +22,14 @@ import org.zowe.jobs.model.Job;
 import org.zowe.jobs.model.JobFile;
 import org.zowe.jobs.model.JobFileContent;
 import org.zowe.jobs.model.JobStatus;
-import org.zowe.jobs.v2.services.JobsServiceV2;
+import org.zowe.jobs.services.JobsServiceV1;
 
 @Slf4j
-@Service
-public class ZosmfJobsServiceV2 implements JobsServiceV2 {
+@Service("ZosmfJobsServiceV1")
+public class ZosmfJobsServiceV1 implements JobsServiceV1 {
 
     @Autowired
-    ZosmfConnectorV2 zosmfConnector;
+    ZosmfConnector zosmfConnector;
 
     // TODO LATER - review error handling, serviceability https://github.com/zowe/jobs/issues/18
     // use the zomsf error categories to work out errors?
@@ -38,49 +38,49 @@ public class ZosmfJobsServiceV2 implements JobsServiceV2 {
     @Override
     public ItemsWrapper<Job> getJobs(String prefix, String owner, JobStatus status) throws ZoweApiException {
         GetJobsZosmfRequestRunner runner = new GetJobsZosmfRequestRunner(prefix, owner, status);
-        return runner.run(zosmfConnector);
+        return runner.runWithLtpa(zosmfConnector);
     }
 
     @Override
     public Job getJob(String jobName, String jobId) {
         GetJobZosmfRequestRunner runner = new GetJobZosmfRequestRunner(jobName, jobId);
-        return runner.run(zosmfConnector);
+        return runner.runWithLtpa(zosmfConnector);
     }
 
     @Override
     public Job submitJobString(String jcl) {
         SubmitJobStringZosmfRequestRunner runner = new SubmitJobStringZosmfRequestRunner(jcl);
-        return runner.run(zosmfConnector);
+        return runner.runWithLtpa(zosmfConnector);
     }
 
     @Override
     public Job submitJobFile(String fileName) {
         SubmitJobFileZosmfRequestRunner runner = new SubmitJobFileZosmfRequestRunner(fileName);
-        return runner.run(zosmfConnector);
+        return runner.runWithLtpa(zosmfConnector);
     }
 
     @Override
     public void purgeJob(String jobName, String jobId) {
         PurgeJobZosmfRequestRunner runner = new PurgeJobZosmfRequestRunner(jobName, jobId);
-        runner.run(zosmfConnector);
+        runner.runWithLtpa(zosmfConnector);
     }
     
     @Override
     public void modifyJob(String jobName, String jobId, String command) {
         ModifyJobZosmfRequestRunner runner = new ModifyJobZosmfRequestRunner(jobName, jobId, command);
-        runner.run(zosmfConnector);
+        runner.runWithLtpa(zosmfConnector);
     }
 
     @Override
     public ItemsWrapper<JobFile> getJobFiles(String jobName, String jobId) {
         GetJobFilesZosmfRequestRunner runner = new GetJobFilesZosmfRequestRunner(jobName, jobId);
-        return runner.run(zosmfConnector);
+        return runner.runWithLtpa(zosmfConnector);
     }
 
     @Override
     public JobFileContent getJobFileContent(String jobName, String jobId, String fileId) {
         GetJobFileContentZosmfRequestRunner runner = new GetJobFileContentZosmfRequestRunner(jobName, jobId, fileId);
-        return runner.run(zosmfConnector);
+        return runner.runWithLtpa(zosmfConnector);
     }
 
     @Override
